# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PROSPECTIVE Web API",
    "description": "REST + WebSocket API for cerebral aneurysm segmentation, detection, morphometry, perforator risk assessment and stent planning.\n\nAll linear dimensions are in **millimetres (mm)**. All HU values follow the standard Hounsfield scale.",
    "contact": {
      "name": "SkullApp — Laboratorio de Imagen Médica",
      "email": "ingprospective@skullapp.tech"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/devices/{session_id}": {
      "get": {
        "tags": [
          "devices"
        ],
        "summary": "Which devices are currently placed",
        "description": "Lists the device kinds with a saved plan in this session. The devices panel uses it to enable its «Limpiar» actions after a session is resumed, when the browser has no memory of what was placed.",
        "operationId": "list_placed_devices_api_devices__session_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceClearResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "devices"
        ],
        "summary": "Remove placed devices from the plan",
        "description": "Deletes the placed device meshes and the state records that feed the PDF report and the DICOM SR, so another device can be planned on a clean scene. `kind` clears one family (`clips`, `coils` or `stent` — the latter covers both the straight catalogue stent and the centreline-guided one); omit it to clear every device.\n\nIdempotent: clearing a device that was never placed succeeds and reports nothing removed.",
        "operationId": "clear_devices_endpoint_api_devices__session_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Device family to clear: 'clips', 'coils' or 'stent'. Omit for all.",
              "title": "Kind"
            },
            "description": "Device family to clear: 'clips', 'coils' or 'stent'. Omit for all."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceClearResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeviceClearResult": {
        "properties": {
          "cleared": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Cleared",
            "description": "Device families this call cleared ('clips' | 'coils' | 'stent')"
          },
          "remaining": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Remaining",
            "description": "Device families that still hold a plan after the call"
          },
          "meshes_removed": {
            "type": "integer",
            "title": "Meshes Removed",
            "description": "Device mesh files (.vtp) deleted from the session",
            "default": 0
          },
          "mesh_urls": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Mesh Urls",
            "description": "Mesh URL per still-placed device family. Lets a resumed session put its devices back in the 3D viewer, which otherwise showed an empty scene while the report still listed them."
          }
        },
        "type": "object",
        "title": "DeviceClearResult",
        "description": "Outcome of removing placed devices from a session's plan."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}